add plot, change structure

+++ ants/models/Ant/Ant.cpp
@@ -1,21 +1,24 @@
 #include "Ant.h"
-#include "PriorityQueue/PriorityQueue.h"
+#include "../PriorityQueue/PriorityQueue.h"
 #include <cmath>
 #include <random>
 #include <chrono>
 
+// Конструктор
 Ant::Ant(Node* startPosition) {
     this->startPosition = startPosition;
     this->currentPosition = startPosition;
 }
 
-Node* Ant::chooseWay(const int alpha, const int beta) {
-    node_iterator nb_it = currentPosition->nb_begin();
-    node_iterator nb_end = currentPosition->nb_end();
-    PriorityQueue probalitiesNodes;
-    bool isStartAvailable = false;
-    double sumProbability = 0.0;
+// Выбрать путь
+Node* Ant::chooseWay(const double alpha, const double beta) {
+    node_iterator nb_it = currentPosition->nb_begin(); // Итератор
+    node_iterator nb_end = currentPosition->nb_end(); // Ссылка на конец
+    PriorityQueue probalitiesNodes; // Очередь с приоритетом по числу p (которое считаем по формуле)
+    bool isStartAvailable = false; // Есть ли стартовая вершина в соседях
+    double sumProbability = 0.0; // Сумма вероятностей
     for (node_iterator it = nb_it; it != nb_end; it++) {
+        // Пропускаем уже посещенные узлы
         if (visitedNodes.find(*it) != visitedNodes.end()) {
             if (*it == startPosition) {
                 isStartAvailable = true;
@@ -30,6 +33,7 @@ Node* Ant::chooseWay(const int alpha, const int beta) {
         sumProbability += prob; // Считаем сумму вероятностей для формулы (знаменатель)
     }
 
+    // Если муравей не обошел никакой узел (т.к. ранее все обошел), и есть стартовая позиция в соседях, то идем туда
     if (probalitiesNodes.empty() && isStartAvailable) {
         return startPosition;
     }
@@ -37,6 +41,7 @@ Node* Ant::chooseWay(const int alpha, const int beta) {
     double mainProbability = randomChoice(); // рандомная p
     double cumulative = 0.0; // Кумулятивная сумма для подсчета вероятностей
 
+    // Считаем вероятность
     while (!probalitiesNodes.empty()) {
         MarkedNode markedNode = probalitiesNodes.pop();
         cumulative += markedNode.mark / sumProbability;
@@ -44,7 +49,7 @@ Node* Ant::chooseWay(const int alpha, const int beta) {
             return markedNode.node;
         }
     }
-    return NULL;
+    return NULL; // Если зашел в тупик
 }
 
 // Генератор случайных чисел от 0 до 1
@@ -54,15 +59,18 @@ double Ant::randomChoice() {
     return dist(re);
 }
 
+// Геттер посещенных узлов
 std::set<Node*>* Ant::getVisitedNodes() {
     return &visitedNodes;
 }
 
+// Посетить узел
 void Ant::visitNode(Node* node) {
     currentPosition = node;
     visitedNodes.insert(node);
 }
 
+// Геттер текущего узла
 Node* Ant::getCurrentPosition() {
     return currentPosition;
 }
